Update files based on code review
