+++ packages/peregrine/lib/util/getOutOfStockIndexes.js
@@ -1,3 +1,8 @@
+/**
+ * Find the value_index of out of stock variants
+ * @return {Array} indexes
+ */
+
 export const getOutOfStockIndexes = items => {
     const OUT_OF_STOCK_CODE = 'OUT_OF_STOCK';
     return items
